+++ src/main/java/com/nhnacademy/frontend/admin/controller/AdminBookController.java
@@ -1,9 +1,9 @@
 package com.nhnacademy.frontend.admin.controller;
 
-import com.nhnacademy.frontend.admin.domain.request.BookCreateRequestDto;
-import com.nhnacademy.frontend.admin.domain.response.BookDetailResponseDto;
-import com.nhnacademy.frontend.admin.domain.response.BookResponseDto;
-import com.nhnacademy.frontend.admin.domain.request.BookUpdateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.request.BookCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.response.BookDetailResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.response.BookResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.request.BookUpdateRequestDto;
 import com.nhnacademy.frontend.book.service.BookService;
 import com.nhnacademy.frontend.book.domain.response.SimpleBookResponseDto;
 import com.nhnacademy.frontend.common.exception.ValidationFailedException;
