+++ src/main/java/com/nhnacademy/frontend/admin/controller/BookCategoryController.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.frontend.admin.controller;
 
-import com.nhnacademy.frontend.admin.domain.request.BookCategoryCreateRequestDto;
-import com.nhnacademy.frontend.admin.domain.request.BookCategoryUpdateRequestDto;
-import com.nhnacademy.frontend.admin.domain.response.BookCategoryResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.request.BookCategoryCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.request.BookCategoryUpdateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.response.BookCategoryResponseDto;
 import com.nhnacademy.frontend.book.service.BookService;
 import com.nhnacademy.frontend.common.exception.ValidationFailedException;
 import jakarta.validation.Valid;
