+++ src/main/java/com/nhnacademy/frontend/book/controller/BookController.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.frontend.book.controller;
 
-import com.nhnacademy.frontend.admin.domain.response.BookDetailResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.response.BookDetailResponseDto;
 import com.nhnacademy.frontend.book.service.BookService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
