+++ src/main/java/com/nhnacademy/frontend/book/service/BookService.java
@@ -1,9 +1,9 @@
 package com.nhnacademy.frontend.book.service;
 
-import com.nhnacademy.frontend.admin.domain.request.*;
-import com.nhnacademy.frontend.admin.domain.response.*;
 import com.nhnacademy.frontend.book.domain.response.BookCategoryNodeResponseDto;
 import com.nhnacademy.frontend.book.domain.response.SimpleBookResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.request.*;
+import com.nhnacademy.frontend.common.adapter.dto.response.*;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
