+++ src/main/java/com/nhnacademy/frontend/common/adapter/BookAdapter.java
@@ -1,10 +1,10 @@
 package com.nhnacademy.frontend.common.adapter;
 
-import com.nhnacademy.frontend.admin.domain.request.*;
-import com.nhnacademy.frontend.admin.domain.response.*;
 import com.nhnacademy.frontend.book.domain.response.BookCategoryNodeResponseDto;
 import com.nhnacademy.frontend.book.domain.response.SimpleBookResponseDto;
 import com.nhnacademy.frontend.cart.dto.response.BookResponse;
+import com.nhnacademy.frontend.common.adapter.dto.request.*;
+import com.nhnacademy.frontend.common.adapter.dto.response.*;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.data.domain.Page;
