+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/request/BookCategoryUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.admin.domain.request;
+package com.nhnacademy.frontend.common.adapter.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
