+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/request/BookTagMapCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.admin.domain.request;
+package com.nhnacademy.frontend.common.adapter.dto.request;
 
 import lombok.Data;
