+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/response/BookTagMapResponseDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.admin.domain.response;
+package com.nhnacademy.frontend.common.adapter.dto.response;
 
 import java.util.List;
