+++ src/test/java/com/nhnacademy/frontend/book/controller/BookTagControllerTest.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.frontend.book.controller;
 
 import com.nhnacademy.frontend.admin.controller.BookTagController;
-import com.nhnacademy.frontend.admin.domain.request.BookTagCreateRequestDto;
-import com.nhnacademy.frontend.admin.domain.response.BookTagResponseDto;
+import com.nhnacademy.frontend.common.adapter.dto.request.BookTagCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.response.BookTagResponseDto;
 import com.nhnacademy.frontend.book.service.BookService;
 import com.nhnacademy.frontend.auth.filter.JwtAuthenticationFilter;
 import org.junit.jupiter.api.Test;
