+++ src/test/java/com/nhnacademy/frontend/book/service/BookServiceImplTest.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.frontend.book.service;
 
 import com.nhnacademy.frontend.common.adapter.BookAdapter;
-import com.nhnacademy.frontend.admin.domain.request.*;
-import com.nhnacademy.frontend.admin.domain.response.*;
+import com.nhnacademy.frontend.common.adapter.dto.request.*;
+import com.nhnacademy.frontend.common.adapter.dto.response.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
